Revert "Feat linux websocket"
