+++ libs/http/src/lib.rs
@@ -2,5 +2,3 @@
  pub mod utils;
  pub mod server;
  pub mod websocket;
- #[cfg(target_os = "linux")]
- pub mod client;
