--- libs/net/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "makepad-net"
-version = "0.1.0"
-authors = ["Makepad <info@makepad.nl>"]
-edition = "2021"
-description = "Makepad http utils"
-license = "MIT OR Apache-2.0"
-
-[dependencies]
-openssl = "0.10.64"
-makepad-url = { path = "../url" }
