--- libs/net/src/tcp/mod.rs
@@ -1,2 +1,0 @@
-mod socket;
-pub use socket::Socket;
