--- libs/url/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "makepad-url"
-version = "0.1.0"
-authors = ["Makepad <info@makepad.nl>"]
-edition = "2021"
-description = "Makepad url utils"
-license = "MIT OR Apache-2.0"
