+++ platform/src/os/linux/mod.rs
@@ -10,11 +10,6 @@ pub mod gl_sys;
 pub mod libc_sys;
 pub mod opengl;
 
-#[cfg(target_os="linux")]
-pub mod url_session;
-#[cfg(target_os="linux")]
-pub mod messaging;
-
 #[cfg(not(target_os="android"))]
 pub mod dma_buf;
 #[cfg(not(target_os="android"))]
@@ -60,7 +55,7 @@ pub(crate) use self::alsa_midi::{OsMidiInput, OsMidiOutput};
 pub(crate) use self::android::android_midi::{OsMidiInput, OsMidiOutput};
 
 #[cfg(not(target_os="android"))]
-pub (crate) use messaging::OsWebSocket;
+pub (crate) use web_socket::OsWebSocket;
 
 #[cfg(target_os="android")]
 pub (crate) use self::android::android_web_socket::OsWebSocket;
