+++ platform/src/os/linux/x11/linux_x11.rs
@@ -18,13 +18,12 @@ use {
         makepad_math::dvec2,
         makepad_live_id::*,
         thread::SignalToUI,
-        event::{Event, NetworkResponseChannel},
+        event::Event,
         pass::CxPassParent,
         cx::{Cx, OsType,LinuxWindowParams}, 
         os::cx_stdin::{PollTimers},
         gpu_info::GpuPerformance,
         os::cx_native::EventFlow,
-        url_session::make_http_request
     }
 };
 
@@ -205,7 +204,6 @@ impl Cx {
                 else{
                     self.call_event_handler(&Event::Timer(e))
                 }
-                self.handle_networking_events();
             }
         }
         
@@ -218,13 +216,6 @@ impl Cx {
     }
 
     pub(crate) fn handle_networking_events(&mut self) {
-        let mut out = Vec::new();
-        while let Ok(event) = self.os.network_response.receiver.try_recv() {
-            out.push(event);
-        }
-        if out.len()>0 {
-            self.call_event_handler(&Event::NetworkResponses(out))
-        }        
     }
     
     pub (crate) fn handle_repaint(&mut self, opengl_windows: &mut Vec<OpenglWindow>) {
@@ -341,9 +332,8 @@ impl Cx {
                 },
                 CxOsOp::UpdateMacosMenu(_menu) => {
                 },
-                CxOsOp::HttpRequest {request_id, request} => {
-                    let sender = self.os.network_response.sender.clone();
-                    make_http_request(request_id, request, sender)
+                CxOsOp::HttpRequest{request_id:_, request:_} => {
+                    todo!()
                 },
                 CxOsOp::PrepareVideoPlayback(_, _, _, _, _) => todo!(),
                 CxOsOp::BeginVideoPlayback(_) => todo!(),
@@ -386,7 +376,6 @@ pub struct CxOs {
     pub(crate) media: CxLinuxMedia,
     pub (crate) stdin_timers: PollTimers,
     pub (crate) start_time: Option<Instant>,
-    pub (crate) network_response: NetworkResponseChannel,
     // HACK(eddyb) generalize this to EGL, properly.
     pub(super) opengl_cx: Option<OpenglCx>,
 }
